+++ .vscode/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "web_cafesite1",
+  "version": "1.0.0",
+  "description": "",
+  "main": "main.js",
+  "scripts": {
+    "build": "webpack",
+    "start": "webpack-dev-server"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ug6067ka/web-cafesite1.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ug6067ka/web-cafesite1/issues"
+  },
+  "homepage": "https://github.com/ug6067ka/web-cafesite1#readme"
+}
