+++ src/index.js
@@ -1,0 +1,3 @@
+import { hello } from "./sub.js";
+
+hello();
