+++ src/sub.js
@@ -1,0 +1,3 @@
+export function hello() {
+  alert("Hellow World");
+}
