+++ style.css
@@ -7,23 +7,27 @@ body {
   position: fixed;
   top: 0;
   width: 100%;
-  height: 60px;
-  background-color: rgba(209, 209, 209, 0.8);
+  height: 100px;
   -webkit-transition: all 0.5s ease;
   transition: all 0.5s ease;
   z-index: 999;
+  background: white;
 }
 
 .hidden {
-  background: none;
   -webkit-transition: all 0.5s ease;
   transition: all 0.5s ease;
-  height: 30px;
-  padding-top: 40px;
+  padding-top: 30px;
 }
 
 .hidden a {
-  color: white;
+  color: black;
+}
+
+.hidden a:hover {
+  color: #a5a5a5;
+  text-decoration: inherit;
+  cursor: pointer;
 }
 
 /* other - nav */
@@ -33,6 +37,7 @@ nav {
   letter-spacing: 1px;
   padding: 50px;
   margin: -20px;
+  background-color: transparent;
 }
 
 nav a {
@@ -42,7 +47,7 @@ nav a {
 
 nav li {
   display: inline;
-  margin-right: 16px;
+  margin-right: 44px;
 }
 
 nav .here {
@@ -64,20 +69,6 @@ nav ul {
 }
 
 /* other */
-html,
-body,
-header,
-footer,
-section,
-h1,
-h2,
-ul,
-li,
-pre {
-  margin: 0;
-  padding: 0;
-}
-
 a {
   text-decoration: none;
   margin: 0 5px;
@@ -171,12 +162,29 @@ a {
   -webkit-box-align: center;
       -ms-flex-align: center;
           align-items: center;
+  font-size: 33px;
 }
 
-.menupage_logo {
-  font-size: 60px;
-  padding-bottom: 9px;
-  border-bottom: solid 3px #3c3b3b;
+h2 {
+  display: inline-block;
+  position: relative;
+}
+
+h2:before {
+  content: "";
+  position: absolute;
+  left: 50%;
+  bottom: -10px;
+  /*下線の上下位置調整*/
+  display: inline-block;
+  width: 80px;
+  height: 4px;
+  border-radius: 5px;
+  -webkit-transform: translate(-50%);
+          transform: translate(-50%);
+  /*位置調整*/
+  background-color: black;
+  /*下線の色*/
 }
 
 .menupage_contents {
@@ -213,6 +221,10 @@ a {
 
 .product-hot {
   color: red;
+}
+
+.product-ice {
+  color: blue;
 }
 
 .hero_title {
@@ -289,11 +301,11 @@ a {
 }
 
 .gray {
-  width: 88%;
+  width: 75%;
   height: 425px;
   background-color: #abaaa8;
   margin-left: auto;
-  margin-right: 30px;
+  margin-right: 105px;
   position: relative;
 }
 
@@ -413,7 +425,7 @@ a {
 .controls-img {
   overflow: hidden;
   margin: 0 auto;
-  width: calc(100% - 60px);
+  width: calc(100% - 100px);
 }
 
 .controls_inner {
@@ -812,6 +824,10 @@ a {
   color: #fff;
 }
 
+.menulist_inner {
+  padding: 0px 30px;
+}
+
 .accordion-wrap {
   width: 50%;
   padding: 0px 30px;
@@ -858,5 +874,71 @@ jQueryで付け外しを行う
           transform: rotate(0);
   -webkit-transition: 0.5s;
   transition: 0.5s;
+}
+
+.about_us {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  height: 500px;
+}
+
+.Information {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+  -webkit-box-align: end;
+      -ms-flex-align: end;
+          align-items: flex-end;
+  -webkit-box-pack: end;
+      -ms-flex-pack: end;
+          justify-content: flex-end;
+  margin-left: auto;
+  margin-right: 60px;
+}
+
+.imfo-text {
+  margin: 20px 50px;
+}
+
+.footer-inner {
+  margin: auto;
+  width: 92%;
+  height: 150px;
+  border-top: 1px solid #000;
+  border-bottom: 1px solid #000;
+  margin-top: 24px;
+  margin-bottom: 28px;
+}
+
+.footer-inner ul {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+}
+
+.footer-inner li {
+  width: 35px;
+  height: 35px;
+  margin-right: 25px;
+  list-style: none;
 }
 /*# sourceMappingURL=style.css.map */
