+++ style.scss
@@ -9,21 +9,24 @@ body {
   position: fixed;
   top: 0;
   width: 100%;
-  height: 60px;
-  background-color: hsla(0, 0%, 82%, 0.8);
+  height: 100px;
   transition: all 0.5s ease;
   z-index: 999;
+  background: white;
 }
 
 .hidden {
-  background: none;
   transition: all 0.5s ease;
-  height: 30px;
-  padding-top: 40px;
+  padding-top: 30px;
 }
 
 .hidden a {
-  color: white;
+  color: black;
+  &:hover {
+    color: #a5a5a5;
+    text-decoration: inherit;
+    cursor: pointer;
+  }
 }
 
 /* other - nav */
@@ -34,6 +37,7 @@ nav {
   letter-spacing: 1px;
   padding: 50px;
   margin: -20px;
+  background-color: transparent;
 }
 
 nav a {
@@ -43,7 +47,7 @@ nav a {
 
 nav li {
   display: inline;
-  margin-right: 16px;
+  margin-right: 44px;
 }
 
 nav .here {
@@ -59,20 +63,6 @@ nav ul {
 }
 
 /* other */
-
-html,
-body,
-header,
-footer,
-section,
-h1,
-h2,
-ul,
-li,
-pre {
-  margin: 0;
-  padding: 0;
-}
 
 a {
   text-decoration: none;
@@ -163,11 +153,28 @@ a {
   display: flex;
   justify-content: space-around;
   align-items: center;
+  font-size: 33px;
 }
-.menupage_logo {
-  font-size: 60px;
-  padding-bottom: 9px;
-  border-bottom: solid 3px $cblack;
+// .menupage_logo {
+//   font-size: 60px;
+//   padding-bottom: 9px;
+//   border-bottom: solid 3px $cblack;
+// }
+h2 {
+  display: inline-block;
+  position: relative;
+}
+h2:before {
+  content: "";
+  position: absolute;
+  left: 50%;
+  bottom: -10px; /*下線の上下位置調整*/
+  display: inline-block;
+  width: 80px;
+  height: 4px;
+  border-radius: 5px;
+  transform: translate(-50%); /*位置調整*/
+  background-color: black; /*下線の色*/
 }
 
 .menupage_contents {
@@ -196,7 +203,9 @@ a {
 .product-hot {
   color: red;
 }
-
+.product-ice {
+  color: blue;
+}
 .hero {
   &_title {
     position: absolute;
@@ -253,11 +262,11 @@ a {
 }
 
 .gray {
-  width: 88%;
+  width: 75%;
   height: 425px;
   background-color: #abaaa8;
   margin-left: auto;
-  margin-right: 30px;
+  margin-right: 105px;
   position: relative;
 }
 
@@ -363,7 +372,7 @@ a {
 .controls-img {
   overflow: hidden;
   margin: 0 auto;
-  width: calc(100% - 60px);
+  width: calc(100% - 100px);
 }
 
 .controls_inner {
@@ -744,6 +753,10 @@ a {
   color: #fff;
 }
 
+.menulist_inner {
+  padding: 0px 30px;
+}
+
 .accordion-wrap {
   width: 50%;
   padding: 0px 30px;
@@ -791,4 +804,50 @@ jQueryで付け外しを行う
 .active::after {
   transform: rotate(0);
   transition: 0.5s;
+}
+
+//about_us
+
+.about_us {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 500px;
+}
+
+//Information
+.Information {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-end;
+  justify-content: flex-end;
+  margin-left: auto;
+  margin-right: 60px;
+}
+.imfo-text {
+  margin: 20px 50px;
+}
+
+//footer
+
+.footer-inner {
+  margin: auto;
+  width: 92%;
+  height: 150px;
+  border-top: 1px solid #000;
+  border-bottom: 1px solid #000;
+  margin-top: 24px;
+  margin-bottom: 28px;
+}
+
+.footer-inner ul {
+  display: flex;
+  flex-direction: row;
+}
+.footer-inner li {
+  width: 35px;
+  height: 35px;
+  margin-right: 25px;
+  list-style: none;
 }
